Added tests for next2.
I'm trying to use Hypothesis to generate test cases which are consistent
with passing and failing the fail-fast test. It's been a bit of a
headache, but I got there in the end, and I'm pretty confident in the
code as a result. The only issue I might not have picked up without
Hypothesis is the use of ble instead of blt, which I don't think could
ever have cost any runtime. I'd also made some other errors around not
dividing my costs by two, and some assembler mistakes. It does seem to
be taking a lot of space in my tests to define my strategies!