WIP minimal Forth to bootstrap Gigatron Forth
The idea is to write as much of the Gigatron Forth in Forth as possible.
But of course because the Forth is ROM resident, we need an environment
that can cross-compile. Because our way of emitting code for the
Gigatron is calling Python functions, it's easiest if this is in Python.

Of course it still makes sense to develop this as much in Forth as
possible.

This is a work in progress, but it presents a runable environment with
colon definitions etc.