Make bootstrap forth use even more dictionaries
There is a dictionary which is used in interpreted mode,
a target dictionary which is where new definitions are stored. There is
also now a search order of dictionaries, so that you can control which
versions of words are found. By default the interpreter dictionary is
the target dictionary which is the only one searched.