Ignore RAM->ROM mode when compiling threads.
My initial design had three "modes" of execution - with three
implementations of the NEXT routine.

* ROM mode is direct-threaded code: Threads are (in effect) lists of
  pointers to Gigatron code. NEXT does an indirect load through the IP
  variable, and stores the result in W and increments IP (In practice
  the encoding of the "pointers" is part of this routine). It then
  jumps to the code that W points to. This is unchanged. A consequence
  of this is that all threads in ROM need a machinecode prefix before
  the list of pointers. This costs four instructions in each thread (the
  cost of the code: 1, and a far jump to the real-code: 3). Because ROM
  is, well, read-only, the dictionary is irrelevant; no new ROM threads
  can be defined.
* RAM -> ROM mode is for threads in RAM, calling words in ROM (kernel
  words). This is the bit I'm rethinking.
  As designed this would also have been direct threaded code -
  threads would have been lists of ROM addresses. This also requires a
  machinecode prefix before each thread in ROM - so there were two of
  them. The cross compiler knew to skip the RAM->ROM prefix, and the
  RAM->ROM prefix itself skipped the ROM mode prefix. This too was four
  instructions. When compiling a RAM->ROM thread, words would be found
  in the dictionary, and their addresses would be copied from the
  dictionary into the thread.
* RAM -> RAM mode would be for threads in RAM, calling other
  user-defined words. This would be indirect threaded code - threads
  would have been lists of pointers to RAM - and the target of the
  pointer would itself be a pointer to ROM. When compiling a RAM->RAM
  thread, words would be found in the dictionary, but rather than
  copying a pointer, the thread can instead point at the pointer.

What I'm currently thinking of doing is making RAM->ROM mode much less
distinct from RAM->RAM mode. I'd originally assumed that I'd have a
dictionary of kernel words in ROM, with specialised search words which
fell back to it if there were no user-defined words of the right name.
Instead I now think that I will simply copy the dictionary into RAM, and
have it be the start of the RAM-mode dictionary. This will be easier to
get something working (this is still a proof of concept). RAM->ROM mode
can then be indirect threaded code just as RAM->RAM mode is - removing
the need to be able to switch between modes mid-thread, and we can get
rid of the RAM->ROM prefixes on ROM mode threads.

If it turns out to be profitable, the direct threaded RAM->ROM mode can
be brought back later.

This commit does not really do anything other than stop emitting the
extra prefix on ROM threads (saving some space). The addresses are
stored away for future use in a dictionary that can be copied to RAM,
but that is not yet implemented. In general anything other than the ROM
mode is really a future project... but I'm getting closer!