Allow cross-compiler to handle page-crossing.
As the cross-compiler processes definitions it calls asm.py
functions to emit Gigatron instructions.

However we have a rule that threads are not allowed to cross page-
boundaries, not least because each page that holds (the start of) a
Forth word must start with the restart-or-quit trampoline.

When the cross-compiler encounters a colon definition, it doesn't
know how long it will end up being in terms of instructions emitted, and so
whether it will end up crossing the page. Instead we recognise when we
have gone too far and undo what we have already done.

asm.py doesn't currently offer much support for this (although it might
be helpful to move some of this machinery in there), and so in the
code we monkey about with its internal state, in a thoroughly
unhygenic way.

The net result is that instead of calling the asm.py functions
directly, we call wrapped versions, which are capable of undoing, and then
redoing their effect.