Implement /2. Needed tricks to stay on same page!
The words that need to use the right-shift table are best placed in the same page,
so that they can share the code which returns to Forth.
It seemed logical to have 2* on the same page.
All the implementations begin by loading the datastack pointer into Y,X.
So that was moved into the restart-or-quit trampoline for this page.

I'm not sure that this is sufficiently tested,
and the changes to the test suite are not exactly lovely either.